requirements.txt
aiogram>=3.4,<4.0
SQLAlchemy>=2.0
aiosqlite>=0.20.0
apscheduler==3.10.4
python-dotenv>=1.0.1
pydantic>=2.5
openai>=1.40.0
reportlab>=4.2.0
httpx>=0.27.0
python-dateutil>=2.9.0
uvloop; sys_platform != 'win32'
